Extract user event fixes into dedicated module
