+++ resources/gml-identifiers.json
@@ -3,7 +3,7 @@
     "manualRoot": "vendor/GameMaker-Manual",
     "packageName": null,
     "packageVersion": null,
-    "generatedAt": "2026-02-05"
+    "generatedAt": "2026-02-06"
   },
   "identifiers": {
     "abs": {
