+++ src/plugin/src/transforms/feather/apply-feather-fixes.ts
@@ -47,6 +47,7 @@ import { removeDuplicateEnumMembers, sanitizeEnumAssignments } from "./enum-fixe
 import { parseExample } from "./parser-bootstrap.js";
 import { renameReservedIdentifiers } from "./reserved-identifier-renaming.js";
 import { findDuplicateSemicolonRanges, removeDuplicateSemicolons } from "./semicolon-fixes.js";
+import { annotateMissingUserEvents } from "./user-event-fixes.js";
 import {
     attachFeatherFixMetadata,
     createCallExpressionTargetFixDetail,
@@ -14218,173 +14219,8 @@ function annotateVariableStructProperty(property, diagnostic) {
 }
 
 /**
- * Annotates missing user-event constant references in the AST.
- *
- * ORGANIZATION SMELL: All user-event-related functionality (detection, constant
- * insertion, validation) should be extracted into a dedicated module rather than
- * being scattered through this large Feather-fixes file.
- *
- * RECOMMENDATION: Create src/plugin/src/transforms/feather/user-event-fixes.ts and
- * move all user-event-specific logic there:
- *   - annotateMissingUserEvents
- *   - insertUserEventConstant
- *   - validateUserEventConstant
- *   - USER_EVENT_CONSTANTS (if defined)
- *
- * This makes the code easier to navigate and test, and reduces the size of this
- * already-oversized file.
+ * Harmonizes texture pointer ternary expressions that use -1 instead of pointer_null.
  */
-function annotateMissingUserEvents({ ast, diagnostic }) {
-    if (!hasFeatherDiagnosticContext(ast, diagnostic)) {
-        return [];
-    }
-
-    const fixes = [];
-
-    const visit = (node) => {
-        if (!node) {
-            return;
-        }
-
-        if (Array.isArray(node)) {
-            for (const entry of node) {
-                visit(entry);
-            }
-            return;
-        }
-
-        if (typeof node !== "object") {
-            return;
-        }
-
-        if (node.type === "CallExpression") {
-            const fix = annotateUserEventCall(node, diagnostic);
-
-            if (fix) {
-                fixes.push(fix);
-                return;
-            }
-        }
-
-        for (const value of Object.values(node)) {
-            if (value && typeof value === "object") {
-                visit(value);
-            }
-        }
-    };
-
-    visit(ast);
-
-    return fixes;
-}
-
-function annotateUserEventCall(node, diagnostic) {
-    const eventInfo = getUserEventReference(node);
-
-    if (!eventInfo) {
-        return null;
-    }
-
-    const fixDetail = createFeatherFixDetail(diagnostic, {
-        target: eventInfo.name,
-        automatic: false,
-        range: {
-            start: Core.getNodeStartIndex(node),
-            end: Core.getNodeEndIndex(node)
-        }
-    });
-
-    if (!fixDetail) {
-        return null;
-    }
-
-    attachFeatherFixMetadata(node, [fixDetail]);
-
-    return fixDetail;
-}
-
-function createUserEventInfo(argumentNode: GameMakerAstNode) {
-    const eventIndex = resolveUserEventIndex(argumentNode);
-
-    if (eventIndex === null) {
-        return null;
-    }
-
-    return { index: eventIndex, name: formatUserEventName(eventIndex) };
-}
-
-function getUserEventReference(node) {
-    if (!node || node.type !== "CallExpression") {
-        return null;
-    }
-
-    const callee = Core.getCallExpressionIdentifier(node);
-    const args = Core.getCallExpressionArguments(node);
-
-    if (Core.isIdentifierWithName(callee, "event_user")) {
-        return createUserEventInfo(args[0]);
-    }
-
-    if (Core.isIdentifierWithName(callee, "event_perform")) {
-        if (args.length < 2 || !Core.isIdentifierWithName(args[0], "ev_user")) {
-            return null;
-        }
-
-        return createUserEventInfo(args[1]);
-    }
-
-    if (Core.isIdentifierWithName(callee, "event_perform_object")) {
-        if (args.length < 3) {
-            return null;
-        }
-
-        return createUserEventInfo(args[2]);
-    }
-
-    return null;
-}
-
-function resolveUserEventIndex(node) {
-    if (!node) {
-        return null;
-    }
-
-    if (node.type === "Literal") {
-        const numericValue = typeof node.value === "number" ? node.value : Number(node.value);
-
-        if (!Number.isInteger(numericValue) || numericValue < 0 || numericValue > 15) {
-            return null;
-        }
-
-        return numericValue;
-    }
-
-    if (node.type === "Identifier") {
-        const match = /^ev_user(\d+)$/.exec(node.name);
-
-        if (!match) {
-            return null;
-        }
-
-        const numericValue = Number.parseInt(match[1]);
-
-        if (!Number.isInteger(numericValue) || numericValue < 0 || numericValue > 15) {
-            return null;
-        }
-
-        return numericValue;
-    }
-
-    return null;
-}
-
-function formatUserEventName(index) {
-    if (!Number.isInteger(index)) {
-        return null;
-    }
-
-    return `User Event ${index}`;
-}
 function harmonizeTexturePointerTernary(node, parent, property, diagnostic) {
     if (!node || node.type !== "TernaryExpression") {
         return null;
