+++ src/plugin/src/transforms/feather/index.ts
@@ -5,3 +5,4 @@ export * from "./ast-traversal.js";
 export * from "./enum-handling.js";
 export * from "./parser-bootstrap.js";
 export * from "./reserved-identifier-renaming.js";
+export * from "./user-event-fixes.js";
